moved the ItemList code into the right file and framed up the AddItemForm component
